@@ -1,119 +1,143 @@
-import { Injectable } from '@angular/core';
-import { HttpClient, HttpErrorResponse } from '@angular/common/http';
-import { catchError, tap } from 'rxjs/operators';
-import { throwError, Subject, BehaviorSubject } from 'rxjs';
-import { User } from './user.model';
-import { Router } from '@angular/router';
+import { Injectable } from "@angular/core";
+import { HttpClient, HttpErrorResponse } from "@angular/common/http";
+import { catchError, tap } from "rxjs/operators";
+import { throwError, Subject, BehaviorSubject } from "rxjs";
+import { User } from "./user.model";
+import { Router } from "@angular/router";
 
-export interface AuthResponseData{
-    kind : string;
-    idToken:string,
-    email:string;
-    refreshToken:string;
-    expiresIn:string;
-    localId : string;
-    resgisterd? : boolean;
+export interface AuthResponseData {
+  kind: string;
+  idToken: string;
+  email: string;
+  refreshToken: string;
+  expiresIn: string;
+  localId: string;
+  resgisterd?: boolean;
 }
-@Injectable({providedIn :'root'})
-export class AuthService{
+@Injectable({ providedIn: "root" })
+export class AuthService {
+  user = new BehaviorSubject<User>(null);
+  private tokenExpirationTimer: any;
+  //token:string  = null;
+  constructor(private http: HttpClient, private router: Router) {}
 
-   user = new BehaviorSubject<User>(null);
-   private tokenExpirationTimer:any;
-   //token:string  = null;
-    constructor(private http:HttpClient,private router:Router){
-
-    }
-    
-  signup(email:string,password:string){
-    return this.http.post<AuthResponseData>('https://identitytoolkit.googleapis.com/v1/accounts:signUp?key=AIzaSyCIMgb4kta0CZpxBmCrq96_5UJg1kqTgR8',
-    {
-       email: email,
-       password:password,
-       returnSecureToken:true
-    }).pipe(catchError(this.handleError),tap(resData =>{
-     this.handleAuthentication(resData.email,resData.localId,resData.idToken,+resData.expiresIn)
-    }));
+  signup(email: string, password: string) {
+    return this.http
+      .post<AuthResponseData>(
+        "https://identitytoolkit.googleapis.com/v1/accounts:signUp?key=AIzaSyBi1PFZvpxlc4Puehu7ecKWPwnFdpnT5G4",
+        {
+          email: email,
+          password: password,
+          returnSecureToken: true,
+        }
+      )
+      .pipe(
+        catchError(this.handleError),
+        tap((resData) => {
+          this.handleAuthentication(
+            resData.email,
+            resData.localId,
+            resData.idToken,
+            +resData.expiresIn
+          );
+        })
+      );
   }
 
-
-  login(email:string,password:string){
-    return this.http.post<AuthResponseData>('https://identitytoolkit.googleapis.com/v1/accounts:signInWithPassword?key=AIzaSyCIMgb4kta0CZpxBmCrq96_5UJg1kqTgR8',
-    {
-      email: email,
-      password:password,
-      returnSecureToken:true
-   }).pipe(catchError(this.handleError),tap(resData =>{
-    this.handleAuthentication(resData.email,resData.localId,resData.idToken,+resData.expiresIn)
-   }));
+  login(email: string, password: string) {
+    return this.http
+      .post<AuthResponseData>(
+        "https://identitytoolkit.googleapis.com/v1/accounts:signInWithPassword?key=AIzaSyCIMgb4kta0CZpxBmCrq96_5UJg1kqTgR8",
+        {
+          email: email,
+          password: password,
+          returnSecureToken: true,
+        }
+      )
+      .pipe(
+        catchError(this.handleError),
+        tap((resData) => {
+          this.handleAuthentication(
+            resData.email,
+            resData.localId,
+            resData.idToken,
+            +resData.expiresIn
+          );
+        })
+      );
   }
 
-  autoLogin(){
-    const userData:{
-      email:string;
+  autoLogin() {
+    const userData: {
+      email: string;
       id: string;
-      _token:string;
-      _tokenExpirationDate:string;
-    } = JSON.parse(localStorage.getItem('userData'));
-    if(!userData){
+      _token: string;
+      _tokenExpirationDate: string;
+    } = JSON.parse(localStorage.getItem("userData"));
+    if (!userData) {
       return;
     }
 
     const LoadedUser = new User(
-      userData.email,userData.id,userData._token,new Date(userData._token)
-      );
-      if(LoadedUser.token){
-        this.user.next(LoadedUser);
-        const expirationDuration = new  Date(userData._tokenExpirationDate).getTime() - new Date().getTime();
-        this.autoLogout(expirationDuration);
-      }
+      userData.email,
+      userData.id,
+      userData._token,
+      new Date(userData._token)
+    );
+    if (LoadedUser.token) {
+      this.user.next(LoadedUser);
+      const expirationDuration =
+        new Date(userData._tokenExpirationDate).getTime() -
+        new Date().getTime();
+      this.autoLogout(expirationDuration);
+    }
   }
-logout(){
-  this.user.next(null);
-  this.router.navigate(['/auth']);
-  localStorage.removeItem('userData');
-  if(this.tokenExpirationTimer){
-    clearTimeout(this.tokenExpirationTimer);
+  logout() {
+    this.user.next(null);
+    this.router.navigate(["/auth"]);
+    localStorage.removeItem("userData");
+    if (this.tokenExpirationTimer) {
+      clearTimeout(this.tokenExpirationTimer);
+    }
+    this.tokenExpirationTimer = null;
   }
-  this.tokenExpirationTimer = null;
-}
 
-autoLogout(expirationDuration : number){
- this.tokenExpirationTimer =  setTimeout(()=>{
-   this.logout();
+  autoLogout(expirationDuration: number) {
+    this.tokenExpirationTimer = setTimeout(() => {
+      this.logout();
+    }, expirationDuration);
+  }
 
- },expirationDuration);
-}
-
-  handleAuthentication(email:string,userId:string ,token:string,expiresIn:number){
+  handleAuthentication(
+    email: string,
+    userId: string,
+    token: string,
+    expiresIn: number
+  ) {
     const expirationDate = new Date(new Date().getTime() + +expiresIn * 1000);
-    const user = new User(
-      email,
-      userId,
-      token,
-      expirationDate);
-      this.user.next(user);
-      console.log(user);
-      this.autoLogout(expiresIn * 1000);
-      localStorage.setItem('userData',JSON.stringify(user));
-      
+    const user = new User(email, userId, token, expirationDate);
+    this.user.next(user);
+    console.log(user);
+    this.autoLogout(expiresIn * 1000);
+    localStorage.setItem("userData", JSON.stringify(user));
   }
 
-  private handleError(errorRes : HttpErrorResponse){
-    let errorMessage = 'An unknown errr occured';
-    if(!errorRes.error || !errorRes.error.error){
+  private handleError(errorRes: HttpErrorResponse) {
+    let errorMessage = "An unknown errr occured";
+    if (!errorRes.error || !errorRes.error.error) {
       return throwError(errorMessage);
     }
-    switch(errorRes.error.error.message){
-      case 'EMAIL_EXISTS' :
-          errorMessage = 'This email exists already';
-          break;
-          case 'EMAIL_NOT_FOUND' :
-            errorMessage = 'This email does not exists';
-            break;
-            case 'INVALID_PASSWORD':
-              errorMessage = 'This password is not correct';
-              break;
-  }
-  return throwError(errorMessage);
+    switch (errorRes.error.error.message) {
+      case "EMAIL_EXISTS":
+        errorMessage = "This email exists already";
+        break;
+      case "EMAIL_NOT_FOUND":
+        errorMessage = "This email does not exists";
+        break;
+      case "INVALID_PASSWORD":
+        errorMessage = "This password is not correct";
+        break;
+    }
+    return throwError(errorMessage);
   }
 }
