@@ -1,16 +1,14 @@
-import { Component, OnInit } from '@angular/core';
-import { AuthService } from './auth/auth.service';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.css']
 })
-export class AppComponent implements OnInit {
-  constructor(private authService:AuthService){
+export class AppComponent {
+  loadedFeature = 'recipe';
 
+  onNavigate(feature: string) {
+    this.loadedFeature = feature;
   }
-  ngOnInit(){
-this.authService.autoLogin();
-   }
 }
