@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Params, Router } from '@angular/router';
 import { FormGroup, FormControl, FormArray, Validators } from '@angular/forms';
+
 import { RecipeService } from '../recipe.service';
-import { Recipe } from '../recipe.model';
 
 @Component({
   selector: 'app-recipe-edit',
@@ -10,51 +10,43 @@ import { Recipe } from '../recipe.model';
   styleUrls: ['./recipe-edit.component.css']
 })
 export class RecipeEditComponent implements OnInit {
-id: number;
-editMode =  false;
-recipeForm : FormGroup;
+  id: number;
+  editMode = false;
+  recipeForm: FormGroup;
 
-  constructor(private route : ActivatedRoute,
-    private recipeService : RecipeService,
-    private router:Router) { }
+  constructor(
+    private route: ActivatedRoute,
+    private recipeService: RecipeService,
+    private router: Router
+  ) {}
 
   ngOnInit() {
-    this.route.params
-    .subscribe(
-      (params:Params)=>{
-        this.id = + params['id'];
-        this.editMode = params['id'] != null;
-        //console.log(this.editMode);
-        this.initForm();
-      }
-    );
+    this.route.params.subscribe((params: Params) => {
+      this.id = +params['id'];
+      this.editMode = params['id'] != null;
+      this.initForm();
+    });
   }
 
-  onSubmit(){
-     //console.log(this.recipeForm);
-    //  const newRecipe = new Recipe(
-    //    this.recipeForm.value['name'],
-    //    this.recipeForm.value['description'],
-    //    this.recipeForm.value['imagePath'],
-    //    this.recipeForm.value['ingredients']);
-     if(this.editMode){
-       this.recipeService.updateRecipe(this.id,this.recipeForm.value);
-     }else{
-       this.recipeService.addRecipe(this.recipeForm.value);
-     }
-     this.onCancel();
+  onSubmit() {
+    // const newRecipe = new Recipe(
+    //   this.recipeForm.value['name'],
+    //   this.recipeForm.value['description'],
+    //   this.recipeForm.value['imagePath'],
+    //   this.recipeForm.value['ingredients']);
+    if (this.editMode) {
+      this.recipeService.updateRecipe(this.id, this.recipeForm.value);
+    } else {
+      this.recipeService.addRecipe(this.recipeForm.value);
+    }
+    this.onCancel();
   }
 
-
-  onDeleteIngredients(index:number){
-  (<FormArray>this.recipeForm.get('ingredients')).removeAt(index);
-  }
-
-  onAddIngredients(){
+  onAddIngredient() {
     (<FormArray>this.recipeForm.get('ingredients')).push(
       new FormGroup({
-        'name' : new FormControl(null,Validators.required),
-        'amount': new FormControl(null,[
+        name: new FormControl(null, Validators.required),
+        amount: new FormControl(null, [
           Validators.required,
           Validators.pattern(/^[1-9]+[0-9]*$/)
         ])
@@ -62,25 +54,31 @@ recipeForm : FormGroup;
     );
   }
 
+  onDeleteIngredient(index: number) {
+    (<FormArray>this.recipeForm.get('ingredients')).removeAt(index);
+  }
 
-  private initForm(){
- //const recipe= this.recipeService.getRecipe(this.id);
+  onCancel() {
+    this.router.navigate(['../'], { relativeTo: this.route });
+  }
+
+  private initForm() {
     let recipeName = '';
-    let recipeDescription = '';
     let recipeImagePath = '';
+    let recipeDescription = '';
     let recipeIngredients = new FormArray([]);
 
-    if(this.editMode){
-      const recipe= this.recipeService.getRecipe(this.id);
-      recipeName =recipe.name;
+    if (this.editMode) {
+      const recipe = this.recipeService.getRecipe(this.id);
+      recipeName = recipe.name;
       recipeImagePath = recipe.imagePath;
       recipeDescription = recipe.description;
-      if(recipe['ingredients']){
-        for(let ingredient of recipe.ingredients){
+      if (recipe['ingredients']) {
+        for (let ingredient of recipe.ingredients) {
           recipeIngredients.push(
             new FormGroup({
-              'name' : new FormControl(ingredient.name,Validators.required),
-              'amount':new FormControl(ingredient.amount,[
+              name: new FormControl(ingredient.name, Validators.required),
+              amount: new FormControl(ingredient.amount, [
                 Validators.required,
                 Validators.pattern(/^[1-9]+[0-9]*$/)
               ])
@@ -90,16 +88,11 @@ recipeForm : FormGroup;
       }
     }
 
-  this.recipeForm = new FormGroup({
-    'name' :  new FormControl(recipeName,Validators.required),
-    'imagePath': new FormControl(recipeImagePath,Validators.required),
-    'description' : new FormControl(recipeDescription,Validators.required),
-    'ingredients': recipeIngredients
-  });
+    this.recipeForm = new FormGroup({
+      name: new FormControl(recipeName, Validators.required),
+      imagePath: new FormControl(recipeImagePath, Validators.required),
+      description: new FormControl(recipeDescription, Validators.required),
+      ingredients: recipeIngredients
+    });
   }
-
-  onCancel(){
-   this.router.navigate(['../'],{relativeTo : this.route})
-  }
-
 }
