@@ -1,7 +1,6 @@
-import { Component, OnInit, Input, EventEmitter, Output } from '@angular/core';
-import { Recipe } from '../../recipe.model';
-import { RecipeService } from '../../recipe.service';
+import { Component, OnInit, Input } from '@angular/core';
 
+import { Recipe } from '../../recipe.model';
 
 @Component({
   selector: 'app-recipe-item',
@@ -9,16 +8,9 @@ import { RecipeService } from '../../recipe.service';
   styleUrls: ['./recipe-item.component.css']
 })
 export class RecipeItemComponent implements OnInit {
- @Input() recipe: Recipe;
- @Input() index: number;
- //@Output() recipeSelected = new EventEmitter<void>();
-  // constructor(private recipeService: RecipeService) { }
+  @Input() recipe: Recipe;
+  @Input() index: number;
 
   ngOnInit() {
   }
-
-  // onSelected(){
-  // //  this.recipeSelected.emit();
-  // this.recipeService.recipeSelected.emit(this.recipe);
-  // }
 }
