@@ -1,48 +1,39 @@
-import { Component, OnInit, EventEmitter, Output, OnDestroy } from '@angular/core';
-import { ActivatedRoute } from '@angular/router';
-import { Router } from '@angular/router'; 
-import { Recipe } from '../recipe.model';
-import { RecipeService } from '../recipe.service';
+import { Component, OnInit, OnDestroy } from '@angular/core';
+import { Router, ActivatedRoute } from '@angular/router';
 import { Subscription } from 'rxjs';
 
-
-
+import { Recipe } from '../recipe.model';
+import { RecipeService } from '../recipe.service';
 
 @Component({
   selector: 'app-recipe-list',
   templateUrl: './recipe-list.component.html',
   styleUrls: ['./recipe-list.component.css']
 })
-export class RecipeListComponent implements OnInit,OnDestroy {
- //@Output() recipeWasSelected = new EventEmitter<Recipe>();
+export class RecipeListComponent implements OnInit, OnDestroy {
   recipes: Recipe[];
-  subscription : Subscription;
+  subscription: Subscription;
 
-  constructor(private recipeService:RecipeService,
-    private route:ActivatedRoute,
-    private router:Router
-    ){
-
+  constructor(private recipeService: RecipeService,
+              private router: Router,
+              private route: ActivatedRoute) {
   }
+
   ngOnInit() {
-   this.subscription = this.recipeService.recipesChanged
-    .subscribe(
-      (recipes:Recipe[]) =>{
-        this.recipes = recipes;
-      }
-    );
+    this.subscription = this.recipeService.recipesChanged
+      .subscribe(
+        (recipes: Recipe[]) => {
+          this.recipes = recipes;
+        }
+      );
     this.recipes = this.recipeService.getRecipes();
   }
 
-  // onRecipeSelected(recipe: Recipe){
-  //   this.recipeWasSelected.emit(recipe);
-  // }
-  onNewRecipe(){
-    this.router.navigate(['new'],{relativeTo:this.route});
-    //this.router.navigate(['../',this.id,'edit'])
+  onNewRecipe() {
+    this.router.navigate(['new'], {relativeTo: this.route});
   }
 
-  ngOnDestroy(){
-  this.subscription.unsubscribe();
+  ngOnDestroy() {
+    this.subscription.unsubscribe();
   }
 }
