@@ -1,69 +1,61 @@
-import { Recipe } from './recipe.model';
-import { EventEmitter, Injectable } from '@angular/core';
-import { Ingredient } from '../shared/ingredient.model';
-import { ShoppingListService } from '../shopping-list/shopping-list.service';
+import { Injectable } from '@angular/core';
 import { Subject } from 'rxjs';
 
+import { Recipe } from './recipe.model';
+import { Ingredient } from '../shared/ingredient.model';
+import { ShoppingListService } from '../shopping-list/shopping-list.service';
+
 @Injectable()
-export class RecipeService{
-recipesChanged = new Subject<Recipe[]>();
-    //recipeSelected = new Subject<Recipe>();
-    // private recipes: Recipe[] = [
-    //     new Recipe(
-    //         'test recipe',
-    //         'simple test recipe',
-    //         'https://cdn.pixabay.com/photo/2017/06/21/22/42/recipe-2428926_960_720.jpg',
-    //         [
-    //             new Ingredient('Meat',1),
-    //             new Ingredient('French Fry',20),
-    //         ]),
-    //     new Recipe(
-    //         'test',
-    //         'simple test',
-    //         'https://cdn.pixabay.com/photo/2017/06/21/22/42/recipe-2428926_960_720.jpg',
-    //         [
-    //             new Ingredient('Buns',2),
-    //             new Ingredient('Meat',1),
-    //         ])
-    //  ]; 
+export class RecipeService {
+  recipesChanged = new Subject<Recipe[]>();
 
-    private recipes:Recipe[] =[];
+  // private recipes: Recipe[] = [
+  //   new Recipe(
+  //     'Tasty Schnitzel',
+  //     'A super-tasty Schnitzel - just awesome!',
+  //     'https://upload.wikimedia.org/wikipedia/commons/7/72/Schnitzel.JPG',
+  //     [new Ingredient('Meat', 1), new Ingredient('French Fries', 20)]
+  //   ),
+  //   new Recipe(
+  //     'Big Fat Burger',
+  //     'What else you need to say?',
+  //     'https://upload.wikimedia.org/wikipedia/commons/b/be/Burger_King_Angus_Bacon_%26_Cheese_Steak_Burger.jpg',
+  //     [new Ingredient('Buns', 2), new Ingredient('Meat', 1)]
+  //   )
+  // ];
+  private recipes: Recipe[] = [];
 
-    
+  constructor(private slService: ShoppingListService) {}
 
-     constructor(private slservice :ShoppingListService){
+  setRecipes(recipes: Recipe[]) {
+    this.recipes = recipes;
+    this.recipesChanged.next(this.recipes.slice());
+  }
 
-     }
+  getRecipes() {
+    return this.recipes.slice();
+  }
 
-     setRecipes(recipes:Recipe[]){
-      this.recipes = recipes;
-      this.recipesChanged.next(this.recipes.slice());
-     }
+  getRecipe(index: number) {
+    return this.recipes[index];
+  }
 
-     getRecipes(){
-         return this.recipes.slice();
-     }
+  addIngredientsToShoppingList(ingredients: Ingredient[]) {
+    this.slService.addIngredients(ingredients);
+  }
 
-     getRecipe(index:number){
-        return this.recipes[index];
-     }
+  addRecipe(recipe: Recipe) {
+    this.recipes.push(recipe);
+    this.recipesChanged.next(this.recipes.slice());
+  }
 
-     addIngredientsToShoppingList(ingredients:Ingredient[]){
-     this.slservice.addIngredients(ingredients);
-     }
+  updateRecipe(index: number, newRecipe: Recipe) {
+    this.recipes[index] = newRecipe;
+    this.recipesChanged.next(this.recipes.slice());
+  }
 
-     addRecipe(recipe:Recipe){
-        this.recipes.push(recipe);
-        this.recipesChanged.next(this.recipes.slice());
-     }
-
-     updateRecipe(index:number,newRecipe:Recipe){
-        this.recipes[index] = newRecipe;
-        this.recipesChanged.next(this.recipes.slice());
-     }
-     deleteRecipe(index:number){
-         this.recipes.splice(index,1);
-         this.recipesChanged.next(this.recipes.slice());
-
-     }
+  deleteRecipe(index: number) {
+    this.recipes.splice(index, 1);
+    this.recipesChanged.next(this.recipes.slice());
+  }
 }
