@@ -1,27 +1,28 @@
 import { Injectable } from '@angular/core';
-import { Resolve, ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
+import {
+  Resolve,
+  ActivatedRouteSnapshot,
+  RouterStateSnapshot
+} from '@angular/router';
+
 import { Recipe } from './recipe.model';
 import { DataStorageService } from '../shared/data-storage.service';
 import { RecipeService } from './recipe.service';
 
+@Injectable({ providedIn: 'root' })
+export class RecipesResolverService implements Resolve<Recipe[]> {
+  constructor(
+    private dataStorageService: DataStorageService,
+    private recipesService: RecipeService
+  ) {}
 
+  resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot) {
+    const recipes = this.recipesService.getRecipes();
 
-@Injectable({providedIn:'root'})
-export class RecipeResolverService implements Resolve<Recipe[]>{
-   constructor(private dataStorageService : DataStorageService,private recipeService : RecipeService){
-      
-   }
-
-
-   resolve(route:ActivatedRouteSnapshot,state :RouterStateSnapshot){
-    const recipe = this.recipeService.getRecipes();
-
-    if(recipe.length ===0){
-        return this.dataStorageService.fetchRecipe();
-    }else{
-        return recipe;
+    if (recipes.length === 0) {
+      return this.dataStorageService.fetchRecipes();
+    } else {
+      return recipes;
     }
-   
-}
-
+  }
 }
