@@ -1,24 +1,15 @@
 import { Component, OnInit } from '@angular/core';
-import { Recipe } from './recipe.model';
-import { RecipeService } from './recipe.service';
 
 @Component({
   selector: 'app-recipes',
   templateUrl: './recipes.component.html',
-  styleUrls: ['./recipes.component.css'],
-  ////providers:[RecipeService]
+  styleUrls: ['./recipes.component.css']
 })
 export class RecipesComponent implements OnInit {
-selectedRecipe :Recipe;
+
   constructor() { }
 
   ngOnInit() {
-  //   this.recipeService.recipeSelected
-  //   .subscribe(
-  //     (recipe : Recipe)=>{
-  //       this.selectedRecipe = recipe;
-  //     }
-  //   );
-   }
+  }
 
 }
