@@ -1,11 +1,12 @@
 import { Directive, HostListener, HostBinding } from '@angular/core';
 
 @Directive({
-    selector: '[appDropdown]'
+  selector: '[appDropdown]'
 })
-export class DropdownDirective{
-    @HostBinding('class.open') isOpen = false;
- @HostListener('click') toogleOpen(){
- this.isOpen = !this.isOpen;
- }
+export class DropdownDirective {
+  @HostBinding('class.open') isOpen = false;
+
+  @HostListener('click') toggleOpen() {
+    this.isOpen = !this.isOpen;
+  }
 }
