@@ -1,35 +1,35 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
+import { Subscription } from 'rxjs';
+
 import { Ingredient } from '../shared/ingredient.model';
 import { ShoppingListService } from './shopping-list.service';
-import { Subscription } from 'rxjs';
 
 @Component({
   selector: 'app-shopping-list',
   templateUrl: './shopping-list.component.html',
   styleUrls: ['./shopping-list.component.css']
 })
-export class ShoppingListComponent implements OnInit , OnDestroy {
- ingredients : Ingredient[];
- private igChangeSub:Subscription;
-  constructor(private slService : ShoppingListService) { }
+export class ShoppingListComponent implements OnInit, OnDestroy {
+  ingredients: Ingredient[];
+  private subscription: Subscription;
+
+  constructor(private slService: ShoppingListService) { }
 
   ngOnInit() {
     this.ingredients = this.slService.getIngredients();
-    this.igChangeSub = this.slService.ingredientsChanged
-    .subscribe(
-      (ingredients:Ingredient[]) => {
-        this.ingredients = ingredients;
-      }
-    );
+    this.subscription = this.slService.ingredientsChanged
+      .subscribe(
+        (ingredients: Ingredient[]) => {
+          this.ingredients = ingredients;
+        }
+      );
   }
 
-  onEditItem(index:number){
+  onEditItem(index: number) {
     this.slService.startedEditing.next(index);
   }
-  ngOnDestroy():void{
-    this.igChangeSub.unsubscribe();
+
+  ngOnDestroy() {
+    this.subscription.unsubscribe();
   }
-
-  
-
 }
